build: add start script (#41)
This PR adds a script to start the angular, angular-experimental and
showcase build/watches in sequence. This is necessary, as
angular-experimental might fail, if angular is in the process of being
built and the showcase might fail if any of the other are in the process
of being built.
This allows us to just run `yarn start` and everything should
continuously work during development.