docs: add output type for docs-app (#214)
In the json file generated by compodoc, the outputs have 'unknown' type,
so in the docs-app every output type is unknown.

So the angular-generator file has been changed to correctly set the
import of the OutputRef and then add the type on outputs on newly added
components/output; the change is not retroactive, so the components have
been manually updated.